Revert "temporary workaround to fix blocking."
This reverts commit 8e41e7bb7c485bac133423394bf5b6f43c9552e1.

This commit removed most of blocking artifacts in enc and dec side ouputs
with cpu-used=6 or higher, but introduced another worse 4x4 artefacts if cpu-used=5 or smaller is
used, which call rd_pick_partition() down from 64x64 block.
Also, because w/o this commit, encoded bitstream can be correctly decoded with unmodified
decoder, so that probably tells that bitstream is already correct but only the
reconstructed frame from both enc and dec is not obtained corectly.